% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_combiner.R
\name{covcomb_compute_log_likelihood}
\alias{covcomb_compute_log_likelihood}
\title{Compute the observed data log-likelihood}
\usage{
covcomb_compute_log_likelihood(psi, partial_covs, var_indices, degrees_freedom)
}
\arguments{
\item{psi}{Current estimate of the combined covariance matrix}

\item{partial_covs}{List of partial covariance matrices}

\item{var_indices}{List of variable indices for each partial covariance matrix}

\item{degrees_freedom}{Degrees of freedom for each partial covariance matrix}
}
\value{
The observed data log-likelihood
}
\description{
Compute the observed data log-likelihood
}
