% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_combiner.R
\name{covcomb_compute_sampling_covariance}
\alias{covcomb_compute_sampling_covariance}
\title{Compute the sampling covariance matrix for the combined covariance matrix}
\usage{
covcomb_compute_sampling_covariance(
  psi,
  partial_covs,
  var_indices,
  degrees_freedom
)
}
\arguments{
\item{psi}{Combined covariance matrix}

\item{partial_covs}{List of partial covariance matrices}

\item{var_indices}{List of variable indices for each partial covariance matrix}

\item{degrees_freedom}{Degrees of freedom for each partial covariance matrix}
}
\value{
Sampling covariance matrix for the combined covariance matrix
}
\description{
Compute the sampling covariance matrix for the combined covariance matrix
}
