% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_combiner.R
\name{covcomb_initialize_psi}
\alias{covcomb_initialize_psi}
\title{Initialize the combined covariance matrix using available variances}
\usage{
covcomb_initialize_psi(partial_covs, var_indices, n_vars)
}
\arguments{
\item{partial_covs}{List of partial covariance matrices}

\item{var_indices}{List of variable indices for each partial covariance matrix}

\item{n_vars}{Total number of variables}
}
\value{
Initialized covariance matrix
}
\description{
Initialize the combined covariance matrix using available variances
}
