% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_germplasm.R
\name{getGenoProtocolFromGermVec}
\alias{getGenoProtocolFromGermVec}
\title{Determine genotyping protocol metadata for a set of accessions}
\usage{
getGenoProtocolFromGermVec(germ_id_vec, brapiConnection, verbose = F)
}
\arguments{
\item{germ_id_vec}{A vector of germplasm DbIds.}

\item{brapiConnection}{A BrAPI connection object, typically from
\code{BrAPI::createBrAPIConnection()}, with a \code{$wizard()} method.}

\item{verbose}{Logical; if \code{FALSE} (default), display purrr progress bar
else print for each \code{germplasmDbId}}
}
\value{
A tibble with one row per germplasm, including genotyping protocol
IDs and names as list columns.
}
\description{
Wrapper for getGenoProtocolFromSingleGerm.
}
\examples{
\dontrun{
brapi_conn <- BrAPI::createBrAPIConnection("wheat-sandbox.triticeaetoolbox.org", is_breedbase = TRUE)

germ_geno_protocols <- getGenoProtocolFromGermVec(
  c("1284387", "1382716", "1415479"), brapi_conn)
germ_geno_protocols
}

}
