% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_germplasm.R
\name{getGermplasmFromSingleTrial}
\alias{getGermplasmFromSingleTrial}
\title{Get germplasm metadata for a single trial via BrAPI}
\usage{
getGermplasmFromSingleTrial(study_id, brapiConnection, verbose = TRUE)
}
\arguments{
\item{study_id}{A single studyDbId to query germplasm for.}

\item{brapiConnection}{A BrAPI connection object, typically from
\code{BrAPI::createBrAPIConnection()},
with \code{$post()} and \code{$get()} methods.}

\item{verbose}{Logical; if \code{TRUE}, print messages about the retrieval
process (direct result vs polling, empty results, etc.).}
}
\value{
A data frame of germplasm metadata for the given trial, with one
row per germplasm. Columns include \code{studyDbId}, \code{germplasmDbId},
\code{germplasmName}, and \code{synonym}. If no result is found, this
function may return \code{NULL} or an empty data frame, depending on how
you choose to handle that case.
}
\description{
Queries the BrAPI \code{/search/germplasm} endpoint for a given trial and
returns a data frame of germplasm accessions associated with that trial
Polling is handled if the search is asynchronous.
}
\examples{
brapiConn <- BrAPI::createBrAPIConnection("wheat-sandbox.triticeaetoolbox.org", is_breedbase = TRUE)

germ_df <- getGermplasmFromSingleTrial("8128", brapiConn)
germ_df

}
