% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_germplasm.R
\name{getGermplasmFromTrialVec}
\alias{getGermplasmFromTrialVec}
\title{Get germplasm metadata for multiple trials}
\usage{
getGermplasmFromTrialVec(study_id_vec, brapiConnection, verbose = F)
}
\arguments{
\item{study_id_vec}{A character vector of studyDbIds to query.}

\item{brapiConnection}{A BrAPI connection object as used in
\code{getGermplasmFromSingleTrial()}.}

\item{verbose}{Logical; passed on to \code{getGermplasmFromSingleTrial()} to
control logging. If FALSE display purrr progress bar}
}
\value{
A data frame obtained by row-binding the results of each trial, with
one row per germplasm per trial
}
\description{
Wrapper around \code{\link{getGermplasmFromSingleTrial}} to retrieve and combine
germplasm metadata for a vector of trial IDs.
}
\examples{
\dontrun{
brapiConn <- BrAPI::createBrAPIConnection("wheat-sandbox.triticeaetoolbox.org", is_breedbase = TRUE)

all_germ <- getGermplasmFromTrialVec(c("8128", "9421"), brapiConn)
all_germ
}

}
