% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_trials.R
\name{getTraitsFromTrialVec}
\alias{getTraitsFromTrialVec}
\title{Retrieve what traits were measured for a set of trials by study IDs}
\usage{
getTraitsFromTrialVec(study_db_id, brapiConnection, namesOrIds = "names")
}
\arguments{
\item{brapiConnection}{A BrAPI connection object, typically created by
\code{BrAPI::createBrAPIConnection()},
with \code{$get()} method available.}

\item{namesOrIds}{A string. If "names" return the names of the traits else
return the trait DB IDs.}

\item{study_id_vec}{A character vector of BrAPI study IDs (studyDbId values)
to query.}
}
\value{
A vector of either trait names or trait DB IDs.
}
\description{
Given a vector of BrAPI study IDs, use the wizard function of a Breedbase
connection to compile a vector of all traits measured in at least one trial
}
\examples{
brapiConn <- BrAPI::createBrAPIConnection("wheat-sandbox.triticeaetoolbox.org", is_breedbase = TRUE)

traits <- getTraitsFromTrialVec(c("8128", "9421"), brapiConn)
traits

}
