% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_germplasm.R
\name{getTrialFromGermplasmVec}
\alias{getTrialFromGermplasmVec}
\title{Get trial metadata for multiple germplasms}
\usage{
getTrialFromGermplasmVec(germplasm_id_vec, brapiConnection, verbose = F)
}
\arguments{
\item{germplasm_id_vec}{A character vector of germplasmDbIds to query.}

\item{brapiConnection}{A BrAPI connection object as used in
\code{getTrialFromSingleGermplasm()}.}

\item{verbose}{Logical; passed on to \code{getTrialFromSingleGermplasm()} to
control logging. If FALSE display purrr progress bar}
}
\value{
A data frame obtained by row-binding the results of each germplasm,
with one row per trial per germplasm
}
\description{
Wrapper around \code{\link{getTrialFromSingleGermplasm}} to retrieve and
combine trial metadata for a vector of germplasm IDs.
}
\examples{
\dontrun{
brapi_conn <- BrAPI::createBrAPIConnection("wheat.triticeaetoolbox.org", is_breedbase = TRUE)

all_trial <- getTrialFromGermplasmVec(c("1284387", "1382716"), brapi_conn)
all_trial
}

}
