% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_germplasm.R
\name{getTrialFromSingleGermplasm}
\alias{getTrialFromSingleGermplasm}
\title{Get trial metadata for all trials in which a single germplasm was evaluated}
\usage{
getTrialFromSingleGermplasm(germplasm_id, brapiConnection, verbose = F)
}
\arguments{
\item{germplasm_id}{A single germplasmDbId to query trials for.}

\item{brapiConnection}{A BrAPI connection object, from
\code{BrAPI::createBrAPIConnection()}.}

\item{verbose}{Logical; if \code{TRUE}, print messages about the retrieval
process.}
}
\value{
A data frame of trial metadata for the given germplasm, with one
row per trial. Columns include \code{germplasm_db_id}, \code{study_db_id}, and
\code{studyName}. If no result is found, not sure what happens.
}
\description{
Queries the BrAPI \code{/search/studies} endpoint for a given germplasm and
returns a data frame of trials associated with that germplasm
}
\examples{
\dontrun{
brapiConn <- BrAPI::createBrAPIConnection("wheat.triticeaetoolbox.org", is_breedbase = TRUE)

trial_df <- getTrialFromSingleGermplasm("1284387", brapiConn)
trial_df
}

}
