% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound_functions.R
\name{make_predictathon_file_structure}
\alias{make_predictathon_file_structure}
\title{Save predictions from a user-provided prediction function into the file
structure needed for the T3 Predictathon 2026}
\usage{
make_predictathon_file_structure(
  prediction_function,
  id_or_name = "ID",
  base_directory = "~"
)
}
\arguments{
\item{prediction_function}{A function that takes as input a study ID and an
indicator of whether this is a CV0 or CV00 prediction task and returns a
list with objects "predictions": a data.frame with two columns:
"germplasmName" and "prediction"; "trials": a vector of the names of the
trials used in training; "accessions": a vector of the names of accessions
used in training}

\item{id_or_name}{A string indicating whether your prediction function needs
the trial to be predicted to be identified by its name (id_or_name="name")
or by its T3/Wheat studyDbId (id_or_name="ID", which is the default)}

\item{base_directory}{A string with the path to the directory where you want
prediction function outputs to be stored}
}
\value{
A string with the path to the base directory in which all prediction
files are stored.
}
\description{
If you have a prediction function that takes a study ID and a CV0 or CV00
indicator, this function will run predictions and return the proper file
structure to submit to the Predictathon folks at T3
}
\examples{
\dontrun{
# This is a dummy prediction function that  creates generic and random values
# for the outputs that a real prediction function would have to create. The
# output is a list.  If your prediction makes a list with the same objects,
# it will work with the "make_predictathon_file_structure" function.

dummy_pred_func <- function(study_id, type_of_cv){
  accession_names <- paste0("focal_accession_", study_id, "_", 1:100)
  predictions <- tibble::tibble(germplasmName=accession_names,
                                prediction=rnorm(100))
  trials <- paste0("training_trial_", type_of_cv, "_", study_id, "_", 1:5)
  accessions <- paste0("training_accession_", study_id, "_", 101:300)
  return(list(predictions=predictions, trials=trials, accessions=accessions))
}

# Give your prediction function to the make_predictathon_file_structure
# function and tell it whether you want trials to be predicted to be identified
# by their name or by the T3/Wheat studyDbId.

base_dir <- make_predictathon_file_structure(dummy_pred_func, id_or_name="ID")
list.files(path = base_dir, recursive = TRUE)
}

}
