% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_combiner.R
\name{matrix_to_vector_idx}
\alias{matrix_to_vector_idx}
\title{Map matrix indices to vector index for covariance parameters}
\usage{
matrix_to_vector_idx(i, j)
}
\arguments{
\item{i}{Row index (0-based)}

\item{j}{Column index (0-based)}
}
\value{
Vector index
}
\description{
Map matrix indices to vector index for covariance parameters
}
