% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_germplasm.R
\name{getGenoProtocolFromSingleGerm}
\alias{getGenoProtocolFromSingleGerm}
\title{Get genotyping protocol metadata for a single germplasm}
\usage{
getGenoProtocolFromSingleGerm(germ_id, brapiConnection, verbose = F)
}
\arguments{
\item{germ_id}{The germplasmDbId for the accession of interest.}

\item{brapiConnection}{A BrAPI connection object, typically from
\code{BrAPI::createBrAPIConnection()}, with a \code{$wizard()} method.}
}
\value{
A tibble with a single row containing
\code{germplasmDbId}, \code{genoProtocolDbId}, and \code{genoProtocolName}.
The genotyping protocol columns are list columns, potentially containing
multiple protocol IDs/names.
}
\description{
Queries the T3 AJAX interface using the \code{$wizard()} method of a
brapiConnection to determine which genotyping protocols have been used for a
specific germplasm
}
\examples{
brapiConn <- BrAPI::createBrAPIConnection("wheat-sandbox.triticeaetoolbox.org", is_breedbase = TRUE)

winner_geno_protocols <- getGenoProtocolFromGermVec("1284387", brapiConn)
winner_geno_protocols

}
